requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4

# API
fastapi==0.111.0
uvicorn==0.30.1

# NLP
nltk==3.9
vaderSentiment==3.3.2
spacy==3.7.5
scikit-learn==1.5.1
transformers==4.42.4
torch>=2.2.0

# Utils
requests==2.32.3
tqdm==4.66.4

# Viz (optional for local EDA/Streamlit)
matplotlib==3.9.0
wordcloud==1.9.3
streamlit==1.36.0

# Dev
jupyter==1.0.0
pytest==8.2.0

Faker
vaderSentiment
